// src/app/models/main/mouth-breathing.model.ts
export class MouthModel {

    id: number;
    clinichistory: any;
    date: string;
    delivery_type: string;
    lactation: string;
    lactation_reason: string;
    feeding_bottle: string;
    feeding_bottle_str: string;
    feeding_bottle_nipple: string;
    feeding_bottle_nipple_str: string;
    baby_pacifier: string;
    baby_pacifier_str: string;
    baby_pacifier_nipple: string;
    baby_pacifier_nipple_str: string;
    digital_suction: string;
    digital_suction_str: string;
    onicofagia: string;
    onicofagia_str: string;
    lip_suction: string;
    lip_suction_str: string;
    another_habit: string;
    another_habit_str: string;
    hoarse: string;
    wake_up_several_times: string;
    sleep_mouth_open: string;
    moves_lot: string;
    babea: string;
    grind_teeth: string;
    wakeup_donot_breathe: string;
    has_enuresis: string;
    talk_asleep: string;
    somnolent: string;
    hyperactive: string;
    difficult_focus: string;
    irritates_easily: string;
    fatigue: string;
    headache: string;
    rinitis: string;
    sick_constantly: string;
    tonsillitis: string;
    otitis: string;
    asthma: string;
    hypertrophy: string;
    broncoespasmos: string;
    adenoid_operated: string;
    diagnosis: string;
    user: any;
    state?: number;
}